refactor(search): organize observed collection helpers
