+++ crates/coral-app/src/search/observed/mod.rs
@@ -2,6 +2,7 @@
 
 mod collector;
 mod publisher;
+mod sensitive;
 mod source_scope;
 mod sqlite_queue;
 mod sqlite_store;
